修复bug
1.修复了一个bug，该bug曾导致二级评论不能正确显示（实际仍存在问题，回复者id字段无法正确赋值）。

+++ smart-fitness-api/src/main/java/cn/foldedj/pojo/entity/Evaluations.java
@@ -25,12 +25,12 @@ public class Evaluations {
     /**
      * 评论者ID
      */
-    private Integer commenterId;
+    private Long commenterId;
 
     /**
      * 回复者ID
      */
-    private Integer replierId;
+    private Long replierId;
 
     /**
      * 内容类型
